Change vehicle rotation shear/flip to three shears
The existing vehicle rotation code produces significant discontinuity
when transitioning across the diagonals. Vehicles wildly change shape
when turning from 30 to 45 (or 45 to 60) degrees. This leads to many
unexpected, and some unavoidable, collisions when navigating around
obstacles, including some paths being impassable despite being real-
istically large enough for the vehicle.

This change should reduce the maximum distance a vehicle tile can jump
in a single turn (+/-15 degrees) to 2 tiles. Reducing this to 1 tile
should be possible but I could not come up with a combination of steps
in the float/int conversions to produce that result.

An effect of this change is that the first turn increment away from an
axis no longer produces horizontal shear for some vehicles. Instead it
produces vertical shear first. This results in vehicles getting longer
instead of wider in that first turn, which allows parallel parking
adjacent to walls that was previously impossible.